requirements.txt
bilby
numpy
scipy
matplotlib
pandas
pyyaml
gwsurrogate